The main screen is almost the last

+++ JackPot/app/src/main/java/com/thisteampl/jackpot/main/userController/UserRelatedFilteringcontents.kt
@@ -4,5 +4,5 @@ package com.thisteampl.jackpot.main.userController
 
 class UserRelatedFilteringcontents (
     var career:String, var emoticon:String,var name:String,
-    var position:String, var region:String, var userIndex:Int
+    var position:String, var region:String, var userIndex:Long
 )
